Move all scmdata stuff into its own module
