+++ src/openscm_calibration/store/__init__.py
@@ -1,0 +1,9 @@
+"""
+Store of results
+"""
+
+from __future__ import annotations
+
+from openscm_calibration.store.base import OptResStore
+
+__all__ = ["OptResStore"]
